[ini, Docker] Updated php to v7.3 and xdebug to v2.7.0
Also commented out zend_extension and extension_dir in php.ini so that we don’t have to update them everytime we update php versions. If we fail to update these we will get a “Call to undefined function mysql_connect() in wp-includes/wp-db.php:1643” error. With this commented out we will not have to worry about this.